Extract ToolExecutor from ChatViewModel
- New ToolExecutor class: all 28 tool definitions + execution logic
- ChatViewModel delegates buildToolDefs() and execute() to ToolExecutor
- Moved: tool handlers, location search, data queries, browser tools,
  memory ops, image gen/analysis, MCP dispatch, helper methods
- ChatViewModel: 1257 → 752 lines (-40%)
- ToolExecutor: 253 lines
- Same behavior, zero UI changes
